ApiFlask/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.2
cssselect==1.2.0
dataclasses-json==0.5.13
dateparser==1.1.8
demjson3==3.0.6
dnspython==2.4.0
et-xmlfile==1.1.0
eventlet==0.33.3
facebook-scraper==0.2.59
facebook-sdk==3.1.0
fake-useragent==1.1.3
firebase-admin==6.2.0
Flask==2.3.2
Flask-Dance==7.0.0
Flask-GoogleAuth==0.4.2
Flask-Login==0.6.2
Flask-OAuth==0.12
Flask-SocketIO==5.3.4
Flask-SSLify==0.1.5
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.239
langsmith==0.0.14
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
oauth2==1.9.0.post1
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parse==1.19.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pyee==8.2.2
PyJWT==2.7.0
pymessenger==0.0.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
URLObject==2.4.3
w3lib==2.1.1
websockets==10.4
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2